elt/extract/requirements.txt
google-cloud-storage==3.4.1
stripe==13.0.1